Address Review Comments 2.0

+++ test/e2e/common/utils.py
@@ -14,27 +14,32 @@
 
 import boto3
 
+
 def application_autoscaling_client():
     return boto3.client("application-autoscaling")
+
 
 def sagemaker_endpoint_register_scalable_target(resource_id):
     target_input = {
         "ServiceNamespace": "sagemaker",
         "ResourceId": resource_id,
-        "ScalableDimension": "sagemaker:variant:DesiredInstanceCount", 
+        "ScalableDimension": "sagemaker:variant:DesiredInstanceCount",
         "MinCapacity": 1,
         "MaxCapacity": 2,
     }
 
-    target_response = application_autoscaling_client().register_scalable_target(**target_input)
+    target_response = application_autoscaling_client().register_scalable_target(
+        **target_input
+    )
     return target_response
+
 
 def sagemaker_endpoint_put_scaling_policy(resource_id, policy_name):
     policy_input = {
         "PolicyName": policy_name,
         "ServiceNamespace": "sagemaker",
         "ResourceId": resource_id,
-        "ScalableDimension": "sagemaker:variant:DesiredInstanceCount", 
+        "ScalableDimension": "sagemaker:variant:DesiredInstanceCount",
         "PolicyType": "TargetTrackingScaling",
         "TargetTrackingScalingPolicyConfiguration": {
             "TargetValue": 70.0,
@@ -43,29 +48,66 @@ def sagemaker_endpoint_put_scaling_policy(resource_id, policy_name):
             "PredefinedMetricSpecification": {
                 "PredefinedMetricType": "SageMakerVariantInvocationsPerInstance",
             },
-        }
+        },
     }
 
-    policy_response = application_autoscaling_client().put_scaling_policy(**policy_input)
+    policy_response = application_autoscaling_client().put_scaling_policy(
+        **policy_input
+    )
     return policy_response
+
 
 def sagemaker_endpoint_deregister_scalable_target(resource_id):
     target_input = {
         "ServiceNamespace": "sagemaker",
         "ResourceId": resource_id,
-        "ScalableDimension": "sagemaker:variant:DesiredInstanceCount", 
+        "ScalableDimension": "sagemaker:variant:DesiredInstanceCount",
     }
 
-    target_response = application_autoscaling_client().deregister_scalable_target(**target_input)
+    describe_response = sagemaker_endpoint_describe_scalable_target(resource_id)
+    if len(describe_response["ScalableTargets"]) > 0:
+        application_autoscaling_client().deregister_scalable_target(**target_input)
+
+
+def sagemaker_endpoint_describe_scalable_target(resource_id):
+    target_input = {
+        "ServiceNamespace": "sagemaker",
+        "ResourceIds": [
+            resource_id,
+        ],
+        "ScalableDimension": "sagemaker:variant:DesiredInstanceCount",
+    }
+
+    target_response = application_autoscaling_client().describe_scalable_targets(
+        **target_input
+    )
     return target_response
+
 
 def sagemaker_endpoint_delete_scaling_policy(resource_id, policy_name):
     policy_input = {
         "ServiceNamespace": "sagemaker",
         "ResourceId": resource_id,
-        "ScalableDimension": "sagemaker:variant:DesiredInstanceCount", 
-        "PolicyName": policy_name
+        "ScalableDimension": "sagemaker:variant:DesiredInstanceCount",
+        "PolicyName": policy_name,
     }
 
-    policy_response = application_autoscaling_client().delete_scaling_policy(**policy_input)
+    describe_response = sagemaker_endpoint_describe_scaling_policy(
+        resource_id, policy_name
+    )
+    if len(describe_response["ScalingPolicies"]) > 0:
+        application_autoscaling_client().delete_scaling_policy(**policy_input)
+
+
+def sagemaker_endpoint_describe_scaling_policy(resource_id, policy_name):
+    policy_input = {
+        "ServiceNamespace": "sagemaker",
+        "ResourceId": resource_id,
+        "ScalableDimension": "sagemaker:variant:DesiredInstanceCount",
+        "PolicyNames": [policy_name],
+    }
+
+    policy_response = application_autoscaling_client().describe_scaling_policies(
+        **policy_input
+    )
     return policy_response
